Display in two channels
The result of the segmentation is displayed in the second channel (they are merged after processing, before the show command).
I've changed the Volume type from float to byte, what speeds up the processing and facilitates merging of channels. Is there any reason why we have to work on float?